% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{covidConfirmedMx}
\alias{covidConfirmedMx}
\title{Get official record of confirmed Covid-19 cases in Mexico}
\usage{
covidConfirmedMx()
}
\value{
A `tibble` with 9 columns:
Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited,
Date of Arrival to Mexico, and Date when the case officialy registered.
}
\description{
Get official record of confirmed Covid-19 cases in Mexico. This is a wrapper for \code{\link{getData}} that downloads the most recent
official report (Mexico's Ministry of Health) of Covid-19 cases in the country.
}
