% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{covidOfficialMx}
\alias{covidOfficialMx}
\title{Get official report on Covid-19 cases in Mexico}
\usage{
covidOfficialMx()
}
\description{
Get latest official report on Covid-19 cases in Mexico generated by the Federal Ministry of Health. This is a wrapper for \code{\link{getData}}.
}
