% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{covidSuspectsMx}
\alias{covidSuspectsMx}
\title{Get official record of suspected Covid-19 cases in Mexico}
\usage{
covidSuspectsMx()
}
\value{
A `tibble` with 8 columns:
Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited, and
Date of Arrival to Mexico.
}
\description{
Get official record of suspected Covid-19 cases in Mexico. This is a wrapper for \code{\link{getData}} that downloads the most recent
official report (Mexico's Ministry of Health) of suspected Covid-19 cases in the country.
}
