% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{covidWWSituation}
\alias{covidWWSituation}
\title{Get last worldwide Covid-19 situation report}
\usage{
covidWWSituation()
}
\description{
Get last Covid-19 world wide situation report. This is a wrapper for \code{\link{getData}} that downloads the most recent
situation report generated by the European Centre for Disease Prevention and Control.
}
