% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get official data about Covid-19 cases in Mexico and the World}
\usage{
getData(where, type, date, source, neat)
}
\arguments{
\item{where}{Scope of the report (Mexico or worldWide). \code{character} vector of length 1.}

\item{type}{Get confirmed (\code{confirmed}) or suspect (\code{suspects}) cases?. \code{character} vector of length 1.}

\item{date}{Date (version) of the published results. \code{character} vector of length 1 or \code{Date} object.}

\item{source}{Data source. \code{character} vector of length 1. Possible data sources are:
\code{Serendipia} (Mexico: confirmed, suspects), \code{Guzmart} (Mexico: confirmed), \code{JHU} (World Wide: confirmed),
and \code{ECDC} (World Wide: confirmed).}

\item{neat}{Should data be cleaned (dates, state name, column names)? \code{logical} vector of length 1.}
}
\value{
A \code{tibble} with the requested data.
}
\description{
Get official data about Covid-19 cases. User defines the report scope (Mexico or worldwide),
the type of report (confirmed or suspected cases), the version of the report (i.e. the date of publication),
and the data source. Please see details.
}
\details{
Main data sources are Mexico's Federal Government Ministry of Health (Secretaría de Salud),
the European Centre for Disease Prevention and Control, and Johns Hopkins University Center for Systems Science and
Engineering. In the case of Mexico, although the main data source is the Federal Government Ministry of Health, there
are third parties that generate their own data by auditing and cleaning the official report.
Due to the wide variety of data sources and formats, not all parameter combinations are possible.
Please consider the following table. Each column is a valid parameter value given the data scope.

\tabular{lcccc}{
Where (scope) \tab Source \tab Type \tab Date \tab Neat \cr
\code{Mexico} \tab \code{SSA} \tab \code{confirmed} \tab \code{today} \tab \code{TRUE}, \code{FALSE} \cr
\code{Mexico} \tab \code{Serendipia} \tab \code{confirmed}, \code{suspects} \tab \code{today} or any date from 15/03/2020 onwards. \tab \code{TRUE}, \code{FALSE} \cr
\code{Mexico} \tab \code{Guzmart} \tab \code{confirmed} \tab \code{today} or any date from 16/03/2020 to 06/04/2020 \tab \code{TRUE}, \code{FALSE} \cr
\code{worldWide} \tab \code{ECDC} \tab \code{confirmed} \tab \code{today} or any date from 01/01/2020 onwards. \tab \code{TRUE}, \code{FALSE} \cr
\code{worldWide} \tab \code{JHU} \tab \code{confirmed} \tab \code{today} or any date from 24/01/2020 onwards.\tab \code{TRUE}, \code{FALSE}
}

For example, when requesting data for Mexico, you can either use \code{Serendipia}, \code{Guzmart} or \code{Mexico}
as \code{source} parameters. If you choose \code{Serendipia},
you can only use \code{type="confirmed"} and any date from from 16/03/2020 onwards.

**WARNING: Data from \code{Guzmart} repository stopped daily updates on 06/04/2020. In future versions of this package, this source will be removed.**

This functions is a wrapper for \code{\link{GetFromSerendipia}},
\code{\link{GetFromGuzmart}}, \code{\link{GetFromJHU}}, and \code{\link{GetFromECDC}}.
}
