% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFromECDC}
\alias{GetFromECDC}
\title{Get Covid-19 Data From ECDC}
\usage{

  GetFromECDC(prefixURL = "https://www.ecdc.europa.eu/sites/default/files/documents/COVID-19-geographic-disbtribution-worldwide-",
  date = "today", fileExt = ".xlsx", auth = authenticate(user = ":",
  password = ":", type = "ntlm"), neat = TRUE)
}
\arguments{
\item{prefixURL}{URL prefix for the HTTP request. \code{character} vector of length 1.}

\item{date}{Date (version) of the report. \code{character} vector of length 1 or \code{Date} object.
If \code{character}, date must be in day/month/year format.}

\item{fileExt}{Target file extension in ECDC server \code{character} vector of length 1.}

\item{auth}{Authentication information generated by \code{authenticate}. \code{request} object.}

\item{neat}{Should data be cleaned (dates, state name, column names)? \code{logical} vector of length 1.}
}
\value{
A `tibble` with 9 columns:
Date, Day, Month, Year, New Cases, New Deaths, Country or Territory, Country ID, Country Population (World Bank 2018 Estimates).
}
\description{
Get Covid-19 situation report from the European Centre for Disease Prevention and Control. This situation
report includes new cases and deaths in all countries in the world.
}
