% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFromGuzmart}
\alias{GetFromGuzmart}
\title{Get Covid-19 Data From Katia Guzman's GitHub repo}
\usage{

  GetFromGuzmart(targetURL = "https://github.com/guzmart/covid19_mex/raw/master/01_datos/",
  filePrefix = "covid_mex_", fileExt = ".xlsx", date = "today",
  neat = TRUE)
}
\arguments{
\item{targetURL}{Target URL of the HTTP request. \code{character} vector of length 1.}

\item{filePrefix}{Target file prefix in GitHub repo. \code{character} vector of length 1.}

\item{fileExt}{Target file extension in GitHub repo. \code{character} vector of length 1.}

\item{date}{Date (version) of the report. \code{character} vector of length 1 or \code{Date} object.
#' If \code{character}, date must be in day/month/year format.}

\item{neat}{Should data be cleaned (dates, state name, column names)? \code{logical} vector of length 1.}
}
\value{
A `tibble` with 9 columns:
Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited,
Date of Arrival to Mexico, and Date when case officialy registered.
}
\description{
Get Covid-19 Data From Katia Guzman's GitHub repo
}
\details{
Mexico's Ministry of Health publishes a daily report containing data about
positive and suspected cases of Covid-19 cases in the country (Comunicado Técnico Diario). However, this report is published
as a PDF document, difficulting further analysis. Katia Guzman makes an open format
version of the report that it's also manually checked for errors and includes a column
with the date when the case was first registered in the official count.

**WARNING: Data from \code{Guzmart} repository stopped daily updates on 06/04/2020. In future versions of this package, this source will be removed.**
}
