% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFromJHU}
\alias{GetFromJHU}
\title{Get Covid-19 Data From JHU CSSE}
\usage{

  GetFromJHU(targetURL = "https://github.com/CSSEGISandData/COVID-19/raw/master/csse_covid_19_data/csse_covid_19_daily_reports/",
  fileExt = ".csv", date = "today", neat = TRUE)
}
\arguments{
\item{targetURL}{Target URL of the HTTP request. \code{character} vector of length 1.}

\item{fileExt}{Target file extension in GitHub repo. \code{character} vector of length 1.}

\item{date}{Date (version) of the report. \code{character} vector of length 1 or \code{Date} object.
#' If \code{character}, date must be in day/month/year format.}

\item{neat}{Should data be cleaned (dates, state name, column names)? \code{logical} vector of length 1.}

\item{filePrefix}{Target file prefix in GitHub repo. \code{character} vector of length 1.}
}
\value{
A `tibble` with 12 columns:
FIPS code, City/County, State/Province, COuntry/Region, Date of Report, Lat, Long, Confirmed cases, Deaths,
Healed Cases, Active Cases, Full Geo Key.
}
\description{
Get Covid-19 situation report from the Johns Hopkins University Center for Systems Science and Engineering. This situation
report includes all countries in the world.
}
