% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{parseSerendipia}
\alias{parseSerendipia}
\title{Parse Serendipia's page with SARS-CoV-2 data in Mexico}
\usage{

  parseSerendipia(targetURL = "https://serendipia.digital/2020/03/datos-abiertos-sobre-casos-de-coronavirus-covid-19-en-mexico/",
  targetCSS = "a.wp-block-file__button")
}
\arguments{
\item{targetURL}{Target URL of the HTTP request. `character` vector of length 1.}

\item{targetCSS}{CSS selector of the nodes containing the data files URL. `character` vector of length 1.}
}
\value{
A character vector with the documents URL
}
\description{
Parse Serendipia's page with SARS-CoV-2 data in Mexico.
}
\details{
Mexico's Ministry of Health publishes a daily report containing data about
positive and suspected cases of Covid-19 cases in the country. However, the data is published
as a PDF document, difficulting further analysis. Serendipia, a data based journalism initiative,
publishes CSV and XLSX versions of the official reports. This function makes a request to
Serendipia's Covid-19 data page and parses the links that redirect to the data files.
}
