% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get official data about Covid-19 cases in Mexico}
\usage{
getData(type = "confirmed", date = "today", source = "Guzmart")
}
\arguments{
\item{type}{Get `confirmed` or `suspect` cases?. `character` vector of length 1.}

\item{date}{Date (version) of the published results. `character` vector of length 1 or `date` object.}

\item{source}{Data source. Mexico's Ministry of Health publishes a daily report containing data about
positive and suspected cases of Covid-19 cases in the country.
However, the data is published as a PDF document, difficulting further analysis.
This package relies on third party sources to get an "open data" version of the report.
`character` vector of length 1. Possible data sources are: `Serendipia`.}
}
\value{
A `tibble` with 8 columns:
Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited,
Date of Arrival to Mexico
}
\description{
Get official data about Covid-19 cases in Mexico. Main data source is Mexico's
Federal Government Ministry of Health (Secretaría de Salud). It should be noted, however,
that the open data format of the daily report is generated by third parties, since the original
report is a PDF document.
}
