% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get official data about Covid-19 cases in Mexico}
\usage{
getData(type = "confirmed", date = "today", source = "Serendipia")
}
\description{
Get official data about Covid-19 cases in Mexico
}
