% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFromGuzmart}
\alias{GetFromGuzmart}
\title{Get Covid-19 Data From Katia Guzman's GitHub page}
\usage{

  GetFromGuzmart(targetURL = "https://github.com/guzmart/covid19_mex/raw/master/01_datos/",
  filePrefix = "covid_mex_", fileExt = ".xlsx", date = "today",
  neat = TRUE)
}
\arguments{
\item{targetURL}{Target URL of the HTTP request. `character` vector of length 1.}

\item{filePrefix}{Target file prefix in GitHub repo. `character` vector of length 1.}

\item{fileExt}{Target file extension in GitHub repo. `character` vector of length 1.}

\item{date}{Date (version) of the published results. `character` vector of length 1 or `date` object.}

\item{neat}{Should data be cleaned (dates, state name, column names)? `logical` vector of length 1.}
}
\value{
A `tibble` with 9 columns:
Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited,
Date of Arrival to Mexico, and Date when case officialy registered.
}
\description{
Get Covid-19 Data From Katia Guzman's GitHub page
}
\details{
Mexico's Ministry of Health publishes a daily report containing data about
positive and suspected cases of Covid-19 cases in the country. However, the data is published
as a PDF document, difficulting further analysis. Katia Guzman makes an open format
version of the report that it's also manually checked for errors and includes a column
with the date when the case was first registered in the official count.
}
