% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFromSerendipia}
\alias{GetFromSerendipia}
\title{Get Covid-19 Data From Serendipia's page}
\usage{

  GetFromSerendipia(targetURL = "https://serendipia.digital/2020/03/datos-abiertos-sobre-casos-de-coronavirus-covid-19-en-mexico/",
  targetCSS = "a.wp-block-file__button", type = "confirmed",
  date = "today", neat = TRUE)
}
\arguments{
\item{targetURL}{Target URL of the HTTP request. `character` vector of length 1.}

\item{targetCSS}{CSS selector of the nodes containing the data files URL. `character` vector of length 1.}

\item{type}{Get `confirmed` or `suspect` cases?. `character` vector of length 1.}

\item{date}{Date (version) of the published results. `character` vector of length 1 or `date` object.}

\item{neat}{Should data be cleaned (dates, state name, column names)? `logical` vector of length 1.}
}
\value{
A `tibble` with 8 columns:
Case Number, State, Sex, Age, Symptoms Onset Date, COVID-19 Test Result, Country Visited,
Date of Arrival to Mexico.
}
\description{
Get Covid-19 Data From Serendipia's page
}
\details{
Mexico's Ministry of Health publishes a daily report containing data about
positive and suspected cases of Covid-19 cases in the country. However, the data is published
as a PDF document, difficulting further analysis. Serendipia, a data based journalism initiative,
publishes CSV and XLSX versions of the official reports. This function makes a request to
Serendipia's Covid-19 data page and downloads the specified data table.
}
