% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GetFromSSA}
\alias{GetFromSSA}
\title{Get Covid-19 Data From Mexico's Ministry of Health Open Data Node}
\usage{

  GetFromSSA(targetURL = "http://187.191.75.115/gobmx/salud/datos_abiertos/datos_abiertos_covid19.zip",
  filePrefix = "covid_mex_", fileExt = ".xlsx", date = "today",
  neat = TRUE)
}
\arguments{
\item{targetURL}{Target URL of the HTTP request. \code{character} vector of length 1.}

\item{filePrefix}{Target file prefix in GitHub repo. \code{character} vector of length 1.}

\item{fileExt}{Target file extension in GitHub repo. \code{character} vector of length 1.}

\item{date}{Date (version) of the report. \code{character} vector of length 1 or \code{Date} object.
#' If \code{character}, date must be in day/month/year format.}

\item{neat}{Should data be cleaned (dates, state name, column names)? \code{logical} vector of length 1.}
}
\value{
A `tibble` with 35 columns: \cr
1. Report date, \cr
2. Case ID,
3. Type of facility where the case was registered (USMER facility / Non-USMER facility), \cr
4. Institution in charge of the facility where the case was registered (Local Gov, Federal Gov, Red Cross, Army...), \cr
5. State ID (where the the case was registered), \cr
6. Patient's gender, \cr
7. State ID (where the patient was born), \cr
8. State ID (where the patient currently lives), \cr
9. Municipality ID (where the patient currently lives), \cr
10. Type of patient (hospitalized or home care), \cr
11. Hospital admittance date (if applicable), \cr
12. Symptoms onset date, \cr
13. Date of death (if applicable), \cr
14. The patient is/was intubated?, \cr
15. The patient has/had been diagnosed with pneumonia? \cr
16. Patient's age. \cr
17. Is the patient Mexican or alien? \cr
18. Is the patient pregnant? \cr
19. Does the patiend speak an indigenous language? \cr
20. The patient has/had been diagnosed with diabetes? \cr
21. The patient has/had been diagnosed with COPD? \cr
22. The patient has/had been diagnosed with Asthma? \cr
23. The patient has/had been diagnosed with any form of immunosuppression? \cr
24. The patient has/had been diagnosed with hypertension? \cr
25. The patient has/had been diagnosed with any other comorbidities? \cr
26. The patient has/had been diagnosed with obesity? \cr
27. The patient has/had been diagnosed with any CDV? \cr
28. The patient has/had been diagnosed with CKD? \cr
29. The patient smokes? \cr
30. The patient had contact with other SARS-CoV-2 confirmed cases? \cr
31. SARS-CoV-2 test result (positive, negative, pending) \cr
32. Is the patient an immigrant? \cr
33. If immigrant, where did the patient came from? \cr
34. Patient's nationality? \cr
35. The patient is/was in ICU?
}
\description{
Get Covid-19 official report from Mexico's Ministry of Health open data node.
}
\details{
Mexico's Ministry of Health released an open dataset with information about confirmed SARS-CoV-2 cases in the country.
This function makes a request to Mexico's Federal Government open data system to retrieve the dataset.
*Please keep in mind that this dataset, although official, may present some inconsistencies.*
}
