% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{parseSerendipia}
\alias{parseSerendipia}
\title{Parse Serendipia's page with SARS-CoV-2 data in Mexico}
\usage{

  parseSerendipia(targetURL = "https://serendipia.digital/2020/03/datos-abiertos-sobre-casos-de-coronavirus-covid-19-en-mexico/",
  targetClass = "a.wp-block-file__button")
}
\description{
Parse Serendipia's page with SARS-CoV-2 data in Mexico
}
